Ignore preempt and part of memtest
Preempt can never pass because the priority scheduler the professor
asked us to implement can never preempt without a priority change.

The commented memtest portion can never pass because malloc doesn't
know about our deferred page allocation, so it will allocate itself
out of a job. This could be fixed if we added page eviction, but
that's not part of the assignment.